Update labels and clean up chart data code
