+++ todo.txt
@@ -1,6 +1,4 @@
 graph
-- precompute chart data, not on render
-- update labels to include metrics name (eg. "US Totals - Deaths")
 - style up the layout a bit better
 - map escape key to close the graph
